move ResourceForm to its own module
